+++ src/main/java/com/raczkowski/app/dto/PlanPriceDto.java
@@ -3,6 +3,7 @@ package com.raczkowski.app.dto;
 import com.raczkowski.app.enums.CurrencyCode;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Getter
@@ -11,10 +12,10 @@ import java.time.ZonedDateTime;
 @NoArgsConstructor
 @Builder
 public class PlanPriceDto {
-    private Long amount;
+    private BigDecimal amount;
     private CurrencyCode currency;
     private ZonedDateTime updatedAt;
     private ZonedDateTime createdAt;
-    private Long previousPrice;
+    private BigDecimal previousPrice;
     private ZonedDateTime changedAt;
 }
